python_backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
pydantic>=2.7.4
httpx>=0.26.0
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.18
langchain-openai>=0.0.5
beautifulsoup4>=4.12.0
requests>=2.31.0
aiohttp>=3.9.0
tiktoken>=0.5.2
python-multipart>=0.0.9
pypdf2>=3.0.0
pdf2image>=1.16.0
pytesseract>=0.3.8
numpy>=1.24.0
supabase>=1.0.0
jq>=1.4.0
langchain-text-splitters>=0.0.1
crawl4ai
langgraph
langsmith
tavily-python
pypdf2
langchain-tavily
langgraph-supervisor


